% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{hamming}
\alias{hamming}
\title{Compute the Hamming distance between a sequence and a vector of sequences with a limit. Complexity: O(n).}
\usage{
hamming(alpha, beta, limit, num_threads)
}
\description{
Compute the Hamming distance between a sequence and a vector of sequences with a limit. Complexity: O(n).
}
